requirements.txt
fastpi
uvicorn
python-multipart
pillow
torch
transformers